Updated database, api optimalization
I updated the GetGames apis to just fetch the games rather than having to loop over games multiple times to attach values.
Now i just update the votes when users do vote, so when i need to fetch the games, we do a lot less for loops and api calls.